% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drifblim.R
\name{.province_normalize}
\alias{.province_normalize}
\title{Normalize province input}
\usage{
.province_normalize(province, strict = TRUE)
}
\arguments{
\item{province}{Character vector of province names or abbreviations.}
}
\value{
Character vector of normalized province names in uppercase
  (for example, \code{"BRITISH COLUMBIA"}).
}
\description{
Converts province names or abbreviations to standardized
uppercase full province or territory names.
}
\keyword{internal}
