% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drifblim.R
\name{download_metadata}
\alias{download_metadata}
\title{Copy packaged metadata file to disk}
\usage{
download_metadata(out_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{out_dir}{Character. Output directory for the metadata file.
Defaults to \code{file.path(getwd(), "drifloon_output")}.}

\item{overwrite}{Logical. If \code{TRUE}, overwrite an existing output file.
Default is \code{FALSE}.}
}
\value{
Character path to the copied \code{.rds} file (invisibly).
}
\description{
Copies the packaged \code{HLY_station_info.rds} file to a target directory.
If no directory is supplied, the file is copied to the current working
directory.
}
