% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drifblim.R
\name{.province_normalize}
\alias{.province_normalize}
\title{Normalize province input}
\usage{
.province_normalize(province)
}
\arguments{
\item{province}{Character. Province name or abbreviation.}
}
\value{
Character. Standardized province abbreviation.
}
\description{
Converts province names or abbreviations to standardized
two-letter province codes.
}
\keyword{internal}
