% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drifblim.R
\name{load_metadata}
\alias{load_metadata}
\title{Load station metadata}
\usage{
load_metadata()
}
\value{
Data frame of station metadata.
}
\description{
Loads the hourly station metadata included with the package.
}
\details{
The metadata is loaded from the packaged file
\code{data/HLY_station_info.rds}.
}
